rename ec2 references
