+++ discovery/install/install.go
@@ -16,11 +16,12 @@
 package install
 
 import (
+	_ "github.com/prometheus/prometheus/discovery/amazon"       // register amazon
 	_ "github.com/prometheus/prometheus/discovery/azure"        // register azure
 	_ "github.com/prometheus/prometheus/discovery/consul"       // register consul
 	_ "github.com/prometheus/prometheus/discovery/digitalocean" // register digitalocean
 	_ "github.com/prometheus/prometheus/discovery/dns"          // register dns
-	_ "github.com/prometheus/prometheus/discovery/ec2"          // register ec2
+	_ "github.com/prometheus/prometheus/discovery/dockerswarm"  // register dockerswarm
 	_ "github.com/prometheus/prometheus/discovery/eureka"       // register eureka
 	_ "github.com/prometheus/prometheus/discovery/file"         // register file
 	_ "github.com/prometheus/prometheus/discovery/gce"          // register gce
